Move PopoverContent to its own file for reuse
[#163804357]
